pyproject.toml
[tool.poetry]
name = "push-new-reservation-requests"
version = "1.0.0"
description = ""
authors = ["30ème Ciel <admin@30emeciel.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
dotmap = "^1.3.22"
firebase-admin = "^4.4.0"
slack-sdk = "^3.2.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.15.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
